third_party/blink/web_tests/external/wpt/webrtc/RTCRtpTransceiver-stopping.https-expected.txt

This is a testharness.js-based test.
FAIL [audio] Locally stopped transceiver goes from stopping to stopped assert_equals: currentDirection after negotiation expected "stopped" but got "sendonly"
PASS [audio] Locally stopping a transceiver ends the track
PASS [audio] Remotely stopping a transceiver ends the track
FAIL [audio] Remotely stopped transceiver goes directly to stopped assert_equals: direction during negotiation expected "stopped" but got "recvonly"
PASS [audio] Rollback when transceiver is not removed does not end track
PASS [audio] Rollback when removing transceiver does end the track
FAIL [audio] Rollback when removing transceiver makes it stopped assert_equals: currentDirection indicate stopped expected (string) "stopped" but got (object) null
PASS [audio] Glare when transceiver is not removed does not end track
FAIL [video] Locally stopped transceiver goes from stopping to stopped assert_equals: currentDirection after negotiation expected "stopped" but got "sendonly"
PASS [video] Locally stopping a transceiver ends the track
PASS [video] Remotely stopping a transceiver ends the track
FAIL [video] Remotely stopped transceiver goes directly to stopped assert_equals: direction during negotiation expected "stopped" but got "recvonly"
PASS [video] Rollback when transceiver is not removed does not end track
PASS [video] Rollback when removing transceiver does end the track
FAIL [video] Rollback when removing transceiver makes it stopped assert_equals: currentDirection indicate stopped expected (string) "stopped" but got (object) null
PASS [video] Glare when transceiver is not removed does not end track
Harness: the test ran to completion.
